+++ app/routes/connections.py
@@ -7,8 +7,7 @@ from app.dependencies.acapy_clients import client_from_auth
 from app.dependencies.auth import AcaPyAuth, acapy_auth
 from app.models.connections import AcceptInvitation, CreateInvitation
 from shared.log_config import get_logger
-from shared.models.conversion import conn_record_to_connection
-from shared.models.webhook_topics import Connection
+from shared.models.connection_record import Connection, conn_record_to_connection
 
 logger = get_logger(__name__)
