+++ app/routes/oob.py
@@ -8,8 +8,7 @@ from app.dependencies.auth import AcaPyAuth, acapy_auth
 from app.models.oob import AcceptOobInvitation, ConnectToPublicDid, CreateOobInvitation
 from app.util.credentials import strip_protocol_prefix
 from shared.log_config import get_logger
-from shared.models.conversion import conn_record_to_connection
-from shared.models.webhook_topics import Connection
+from shared.models.connection_record import Connection, conn_record_to_connection
 
 logger = get_logger(__name__)
