+++ app/services/verifier/acapy_verifier_v2.py
@@ -17,9 +17,11 @@ from app.models.verifier import (
 )
 from app.services.verifier.acapy_verifier import Verifier
 from shared.log_config import get_logger
-from shared.models.conversion import presentation_record_to_model as record_to_model
+from shared.models.presentation_exchange import PresentationExchange
+from shared.models.presentation_exchange import (
+    presentation_record_to_model as record_to_model,
+)
 from shared.models.protocol import pres_id_no_version
-from shared.models.webhook_topics import PresentationExchange
 
 logger = get_logger(__name__)
