+++ app/tests/e2e/test_credentials.py
@@ -19,7 +19,7 @@ from app.tests.util.trust_registry import register_issuer
 from app.tests.util.webhooks import check_webhook_state
 from app.util.string import random_version
 from shared import RichAsyncClient
-from shared.models.webhook_topics import CredentialExchange
+from shared.models.credential_exchange import CredentialExchange
 
 CREDENTIALS_BASE_PATH = router.prefix
