+++ app/tests/services/verifier/test_acapy_verifier_v2.py
@@ -15,7 +15,7 @@ from app.tests.verifier.utils import (
     indy_proof_request,
     v20_presentation_exchange_records,
 )
-from shared.models.webhook_topics import PresentationExchange
+from shared.models.presentation_exchange import PresentationExchange
 
 
 @pytest.mark.anyio
