+++ app/tests/util/webhooks.py
@@ -3,7 +3,7 @@ from typing import Dict, Optional
 from app.event_handling.sse_listener import SseListener
 from app.util.tenants import get_wallet_id_from_b64encoded_jwt
 from shared import RichAsyncClient
-from shared.models.webhook_topics import CloudApiTopics
+from shared.models.webhook_events import CloudApiTopics
 
 
 def get_wallet_id_from_async_client(client: RichAsyncClient) -> str:
