+++ endorser/services/endorsement_processor.py
@@ -12,7 +12,7 @@ from shared.constants import (
     GOVERNANCE_LABEL,
 )
 from shared.log_config import get_logger
-from shared.models.webhook_topics.base import Endorsement
+from shared.models.endorsement import Endorsement
 from shared.services.redis_service import RedisService
 from shared.util.rich_parsing import parse_with_error_handling
