+++ endorser/tests/test_endorser_processor.py
@@ -200,7 +200,6 @@ async def test_attempt_process_endorsement_x(endorsement_processor_mock):
 
 @pytest.mark.anyio
 async def test_process_endorsement_event_governance(endorsement_processor_mock):
-
     governance = GOVERNANCE_LABEL
     event_dict = {
         "origin": governance,
