+++ shared/models/webhook_events/__init__.py
@@ -1,0 +1,5 @@
+# flake8: noqa
+from shared.models.webhook_events.payloads import *
+from shared.models.webhook_events.topics import *
+
+WEBHOOK_TOPIC_ALL = "ALL_WEBHOOKS"
