+++ webhooks/models/__init__.py
@@ -1,3 +1,6 @@
 # flake8: noqa
 # pylint: disable=unused-import
-from webhooks.models.webhook_models import *
+from webhooks.models.conversions import *
+from webhooks.models.redis_payloads import *
+from webhooks.models.topic_mapping import *
+from webhooks.models.topic_payloads import *
